+++ CHANGES.rst
@@ -4,9 +4,14 @@ Changelog
 3.0.2 (unreleased)
 ------------------
 
+- Use INameChooser for new id when renaming objects.
+  Fix https://github.com/plone/plone.app.dexterity/issues/73
+  [pbauer]
+
 - Allow folderish types as default_page as long as users cannot add content
   to them.
   [pbauer]
+
 - fix removing tags with non-ascii characters in folder_contents
   [petschki]